ai-service/requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.4
Pillow==10.4.0
